+++ bindings/ruby/test/test_spline.rb
@@ -1,9 +1,9 @@
 require 'minitest/autorun'
-require 'base/geometry/spline'
+require 'sisl/spline'
 
 class TC_Geometry_Spline < Minitest::Test
     def test_base
-        v = Types::Base::Geometry::Spline.new(3)
+        v = SISL::Spline.new(3)
         assert_equal 3, v.dimension
         assert_equal 0.1, v.geometric_resolution
         assert_equal 3, v.order
@@ -14,7 +14,7 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_deriv
-        v = Types::Base::Geometry::Spline.new(3)
+        v = SISL::Spline.new(3)
 
         v.interpolate(
 	    [[0, 0, 0], 
@@ -39,10 +39,10 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_concat
-        v1 = Types::Base::Geometry::Spline.new(3)
+        v1 = SISL::Spline.new(3)
         v1.interpolate([0, 0, 0, 1, 2, 3])
 
-        v2 = Types::Base::Geometry::Spline.new(3)
+        v2 = SISL::Spline.new(3)
         v2.interpolate([2, 2, 2, 4, 6, 8])
 
         result = v1.dup
@@ -54,12 +54,12 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_join_singletons
-        v1 = Types::Base::Geometry::Spline.new(3)
+        v1 = SISL::Spline.new(3)
         v1.interpolate([0, 0, 0])
-        v2 = Types::Base::Geometry::Spline.new(3)
+        v2 = SISL::Spline.new(3)
         v2.interpolate([4, 4, 4])
 
-        expected = Types::Base::Geometry::Spline.new(3)
+        expected = SISL::Spline.new(3)
         expected.interpolate([0, 0, 0, 4, 4, 4])
 
         result = v1.dup
@@ -68,9 +68,9 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_join_curve_and_singleton
-        v1 = Types::Base::Geometry::Spline.new(3)
+        v1 = SISL::Spline.new(3)
         v1.interpolate([0, 0, 0])
-        v2 = Types::Base::Geometry::Spline.new(3)
+        v2 = SISL::Spline.new(3)
         v2.interpolate([4, 4, 4, 6, 6, 6])
 
         result = v1.dup
@@ -79,9 +79,9 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_join_singleton_and_curve
-        v1 = Types::Base::Geometry::Spline.new(3)
+        v1 = SISL::Spline.new(3)
         v1.interpolate([0, 0, 0, 4, 4, 4])
-        v2 = Types::Base::Geometry::Spline.new(3)
+        v2 = SISL::Spline.new(3)
         v2.interpolate([6, 6, 6])
 
         result = v1.dup
@@ -90,10 +90,10 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_join
-        v1 = Types::Base::Geometry::Spline.new(3)
+        v1 = SISL::Spline.new(3)
         v1.interpolate([0, 0, 0, 1, 2, 3])
 
-        v2 = Types::Base::Geometry::Spline.new(3)
+        v2 = SISL::Spline.new(3)
         v2.interpolate([4, 4, 4, 4, 6, 8])
 
         result = v1.dup
@@ -104,19 +104,25 @@ class TC_Geometry_Spline < Minitest::Test
     end
 
     def test_distance_to
-        v1 = Types::Base::Geometry::Spline3.new
+        v1 = SISL::Spline3.new
         v1.interpolate([Eigen::Vector3.new(0, 0, 0), Eigen::Vector3.new(1, 1, 1)])
 
         assert(v1.distance_to(Eigen::Vector3.new(0.5, 0.5, 0.5), 0, 1e-9) < 1e-9)
     end
 
     def test_find_one_closest_point
-        v1 = Types::Base::Geometry::Spline3.new
+        v1 = SISL::Spline3.new
         v1.interpolate([Eigen::Vector3.new(0, 0, 0), Eigen::Vector3.new(1, 1, 1)])
 
         test_p = Eigen::Vector3.new(0.5, 0.5, 0.5)
         t = v1.find_one_closest_point(test_p, 0, 1e-9)
         assert v1.get(t).approx?(test_p)
+    end
+
+    def test_it_is_aliased_to_base_geometry_spline_for_backward_compatibility
+        require 'base/geometry/spline'
+        assert_same SISL::Spline, Types::Base::Geometry::Spline
+        assert_same SISL::Spline3, Types::Base::Geometry::Spline3
     end
 end
